requirements.txt
psutil==5.4.7
Adafruit_DHT==1.4.0
paho_mqtt==1.4.0